refactor(repo): admin monitoring 메서드 별도 module 로 분리
Phase 5.3 가 추가한 3 메서드 (getGlobalInstanceSummary / getTopCrashedServers
/ getCrashTrendByHour) 를 McpCatalogRepository 에서 추출 →
McpAdminMonitoringRepository 신규 module.

# 동기
- max-lines 정책 위반 (522 lines > 400) 해소 → 497 lines
- 책임 분리: catalog/instance/user-server 는 user-scope 격리 / admin
  observability 는 전역 aggregate (별도 권한 모델)
- 새 admin observability 메서드 추가 시 단일 file 영향만

# 변경
- data/repositories/mcp-admin-monitoring-repository.ts 신규 (142 lines)
  - GlobalInstanceSummary / TopCrashedServer / CrashTrendBucket export
- data/repositories/mcp-catalog-repository.ts 의 admin 영역 제거 +
  pointer comment 추가
- routes/mcp-admin-monitoring.routes.ts: import 경로 변경
  (McpCatalogRepository → McpAdminMonitoringRepository)
- jest test: admin block 안의 변수명을 catalogRepo + adminRepo 로 분리
  (record... 는 catalog, getGlobal... 등은 admin)

# 검증
- tsc 0 / lint 0 errors (catalog repo 의 417 lines warning 잔존, 17 lines
  초과 — 다음 작업 후보)
- 전체 회귀: 97 suites / 1604 tests / 0 failed (동일)
